// src/app/api/api-nmviajes/models/rq-gallery-get-by-id-ce-request-1.ts
/* tslint:disable */
/* eslint-disable */
import { CeRequestCaller } from './ce-request-caller';
import { RqGalleryGetById } from './rq-gallery-get-by-id';
export interface RqGalleryGetByIdCeRequest1 {
  Caller: CeRequestCaller;
  MuteExceptions: boolean;
  Parameter: RqGalleryGetById;
  TrackingCode: string;
}